Fix landing auth validation import
