+++ web/src/pages/reset-password.astro
@@ -130,7 +130,7 @@ const supportEmail = siteConfig.links.supportEmail;
 
 <script>
   import { createClient } from '@supabase/supabase-js';
-  import { mapAuthError, validatePassword } from '../../../src/utils/authValidation';
+  import { mapAuthError, validatePassword } from '../../../packages/auth-validation/src';
 
   const SUPABASE_URL = import.meta.env.PUBLIC_SUPABASE_URL;
   const SUPABASE_ANON_KEY = import.meta.env.PUBLIC_SUPABASE_ANON_KEY;
